Refactor FormW2 to use DTO and add new service method
The FormW2 entity has been refactored to use a Data Transfer Object (DTO) for handling data. This brings increased type safety and eases the mapping between the domain model and the persistence layer. Additionally, a new method to find all FormW2 by a TaxReturnId was introduced in the service layer and its corresponding endpoint in the Controller. These enhancements will improve the maintainability and usability of the FormW2-related features.