Added tax form retrieval to TaxReturnController
This update added new functionality to the TaxReturnController to retrieve both FormW2 and Form1099 forms for a given Tax return ID. Changes were also made to the JSON annotations in various model classes to adjust the object serialization behavior. 'JsonIdentityInfo' was removed in some classes and 'JsonIgnore' was added in others preventing recursive associations.